Fix test case failures and remove Linux linker dependency.
This commit removes the reliance on the -Wl,-wrap support of the
Linux linker and instead relies on redefining realloc() through
preprocessor macros for the scope of the vector.h inclusion.

Some test cases seem to consistently fail with newer versions of
cmocka/gcc/clang when optimization is set to -O3.
This seems to be due to some change in optimization during the
compilation or linking stage.

This ?should? NOT trigger undefined behavior, as it does not
redefine an existing function, but instead solely relies on
the C preprocessor.

Removing the Linux linker dependency also means that the test
cases should be reproducible on other platforms like macOS.